+++ src/Extracting/ParsesValidationRules.php
@@ -21,36 +21,72 @@ trait ParsesValidationRules
         $validationRules = $this->normaliseRules($validationRules);
 
         $parameters = [];
+        $dependentRules = [];
         foreach ($validationRules as $parameter => $ruleset) {
             if (count($customParameterData) && !isset($customParameterData[$parameter])) {
                 c::debug($this->getMissingCustomDataMessage($parameter));
             }
             $userSpecifiedParameterInfo = $customParameterData[$parameter] ?? [];
 
+            // Make sure the user-specified description comes first (and add full stops where needed).
+            $description = $userSpecifiedParameterInfo['description'] ?? '';
+            if (!empty($description) && !Str::endsWith($description, '.')) {
+                $description .= '.';
+            }
             $parameterData = [
                 'name' => $parameter,
                 'required' => false,
                 'type' => null,
                 'example' => self::$MISSING_VALUE,
-                'description' => '',
+                'description' => $description,
             ];
+            $dependentRules[$parameter] = [];
+
+            // First, parse only "independent" rules
+            foreach ($ruleset as $rule) {
+                $parsed = $this->parseRule($rule, $parameterData, true);
+                if (!$parsed) {
+                    $dependentRules[$parameter][] = $rule;
+                }
+            }
+
+            $parameterData['description'] = trim($parameterData['description']);
+
+            // Set a default type
+            if (is_null($parameterData['type'])) {
+                $parameterData['type'] = 'string';
+            }
+
+            $parameterData['name'] = $parameter;
+            $parameters[$parameter] = $parameterData;
+        }
+
+        // Now parse any "dependent" rules and set examples. At this point, we should know all field's types.
+        foreach ($dependentRules as $parameter => $ruleset) {
+            $parameterData = $parameters[$parameter];
+            $userSpecifiedParameterInfo = $customParameterData[$parameter] ?? [];
+
+            foreach ($ruleset as $rule) {
+                $this->parseRule($rule, $parameterData, false, $parameters);
+            }
 
             // Make sure the user-specified example overwrites others.
             if (isset($userSpecifiedParameterInfo['example'])) {
-                $parameterData['example'] = $userSpecifiedParameterInfo['example'];
+                if ($this->shouldCastUserExample()) {
+                    // Examples in comments are strings, we need to cast them properly
+                    $parameterData['example'] = $this->castToType($userSpecifiedParameterInfo['example'], $parameterData['type'] ?? 'string');
+                } else {
+                    $parameterData['example'] = $userSpecifiedParameterInfo['example'];
+                }
             }
 
-            foreach ($ruleset as $rule) {
-                $this->parseRule($rule, $parameterData);
-            }
-            if ($this->shouldCastUserExample() && isset($userSpecifiedParameterInfo['example'])) {
-                // Examples in comments are strings, we need to cast them properly
-                $parameterData['example'] = $this->castToType($userSpecifiedParameterInfo['example'], $parameterData['type'] ?? 'string');
+            // End descriptions with a full stop
+            $parameterData['description'] = trim($parameterData['description']);
+            if (!empty($parameterData['description']) && !Str::endsWith($parameterData['description'], '.')) {
+                $parameterData['description'] .= '.';
             }
 
-            $parameterData = $this->setParameterExampleAndDescription($parameterData, $userSpecifiedParameterInfo);
-
-            $parameterData['name'] = $parameter;
+            $parameterData = $this->updateParameterExample($parameterData);
             $parameters[$parameter] = $parameterData;
         }
 
@@ -101,14 +137,20 @@ trait ParsesValidationRules
     /**
      * Parse a validation rule and extract a parameter type, description and setter (used to generate an example).
      *
-     * @param $rule
-     * @param $parameterData
+     * If $independentOnly is true, only independent rules will be parsed.
+     * If a rule depends on another parameter (eg gt:field) or needs to know the type of the parameter first (eg:
+     * size:34), it will return false.
      */
-    protected function parseRule($rule, &$parameterData)
+    protected function parseRule($rule, array &$parameterData, bool $independentOnly, array $allParameters = []): bool
     {
         // Convert string rules into rule + arguments (eg "in:1,2" becomes ["in", ["1", "2"]])
         $parsedRule = $this->parseStringRuleIntoRuleAndArguments($rule);
         [$rule, $arguments] = $parsedRule;
+
+        $dependentRules = ['between', 'max', 'min', 'size', 'gt', 'gte', 'lt', 'lte', 'before', 'after', 'before_or_equal', 'after_or_equal'];
+        if ($independentOnly && in_array($rule, $dependentRules)) {
+            return false;
+        }
 
         // Reminders:
         // 1. Append to the description (with a leading space); don't overwrite.
@@ -158,7 +200,7 @@ trait ParsesValidationRules
                 break;
             case 'file':
                 $parameterData['type'] = 'file';
-                $parameterData['description'] .= 'The value must be a file.';
+                $parameterData['description'] .= ' The value must be a file.';
                 $parameterData['setter'] = function () {
                     return $this->generateDummyValue('file');
                 };
@@ -168,32 +210,32 @@ trait ParsesValidationRules
              * Special string types
              */
             case 'alpha':
-                $parameterData['description'] .= "The value must contain only letters. ";
+                $parameterData['description'] .= " The value must contain only letters.";
                 $parameterData['setter'] = function () {
                     return $this->getFaker()->lexify('??????');
                 };
                 break;
             case 'alpha_dash':
-                $parameterData['description'] .= "The value must contain only letters, numbers, dashes and underscores. ";
+                $parameterData['description'] .= " The value must contain only letters, numbers, dashes and underscores.";
                 $parameterData['setter'] = function () {
                     return $this->getFaker()->lexify('???-???_?');
                 };
                 break;
             case 'alpha_num':
-                $parameterData['description'] .= "The value must contain only letters and numbers. ";
+                $parameterData['description'] .= " The value must contain only letters and numbers.";
                 $parameterData['setter'] = function () {
                     return $this->getFaker()->bothify('#?#???#');
                 };
                 break;
             case 'timezone':
                 // Laravel's message merely says "The value must be a valid zone"
-                $parameterData['description'] .= "The value must be a valid time zone, such as <code>Africa/Accra</code>. ";
+                $parameterData['description'] .= " The value must be a valid time zone, such as <code>Africa/Accra</code>.";
                 $parameterData['setter'] = function () {
                     return $this->getFaker()->timezone;
                 };
                 break;
             case 'email':
-                $parameterData['description'] .= $this->getDescription($rule) . ' ';
+                $parameterData['description'] .= ' ' . $this->getDescription($rule);
                 $parameterData['setter'] = function () {
                     return $this->getFaker()->safeEmail;
                 };
@@ -205,10 +247,10 @@ trait ParsesValidationRules
                 };
                 $parameterData['type'] = 'string';
                 // Laravel's message is "The value format is invalid". Ugh.🤮
-                $parameterData['description'] .= "The value must be a valid URL. ";
+                $parameterData['description'] .= " The value must be a valid URL.";
                 break;
             case 'ip':
-                $parameterData['description'] .= $this->getDescription($rule) . ' ';
+                $parameterData['description'] .= ' ' . $this->getDescription($rule);
                 $parameterData['type'] = 'string';
                 $parameterData['setter'] = function () {
                     return $this->getFaker()->ipv4;
@@ -216,78 +258,106 @@ trait ParsesValidationRules
                 break;
             case 'json':
                 $parameterData['type'] = 'string';
-                $parameterData['description'] .= $this->getDescription($rule) . ' ';
+                $parameterData['description'] .= ' ' . $this->getDescription($rule);
                 $parameterData['setter'] = function () {
                     return json_encode([$this->getFaker()->word, $this->getFaker()->word,]);
                 };
                 break;
             case 'date':
                 $parameterData['type'] = 'string';
-                $parameterData['description'] .= $this->getDescription($rule) . ' ';
-                $parameterData['setter'] = function () {
-                    return date('Y-m-d\TH:i:s', time());
-                };
+                $parameterData['description'] .= ' ' . $this->getDescription($rule);
+                $parameterData['setter'] = fn() => date('Y-m-d\TH:i:s', time());
                 break;
             case 'date_format':
                 $parameterData['type'] = 'string';
                 // Laravel description here is "The value must match the format Y-m-d". Not descriptive enough.
-                $parameterData['description'] .= "The value must be a valid date in the format {$arguments[0]} ";
+                $parameterData['description'] .= " The value must be a valid date in the format {$arguments[0]}.";
                 $parameterData['setter'] = function () use ($arguments) {
                     return date($arguments[0], time());
                 };
                 break;
+            case 'after':
+            case 'after_or_equal':
+                $parameterData['type'] = 'string';
+                $parameterData['description'] .= ' ' . $this->getDescription($rule, [':date' => "<code>{$arguments[0]}</code>"]);
+                // TODO introduce the concept of "modifiers", like date_format
+                $startDate = isset($allParameters[$arguments[0]]) ? null : $arguments[0];
+                $parameterData['setter'] = fn() => $this->getFaker()->dateTimeBetween($startDate)->format('Y-m-d');
+                break;
+            case 'before':
+            case 'before_or_equal':
+                $parameterData['type'] = 'string';
+                // The argument can be either another field or a date
+                $endDate = isset($allParameters[$arguments[0]]) ? null : $arguments[0];
+                $parameterData['description'] .= ' ' . $this->getDescription($rule, [':date' => "<code>{$arguments[0]}</code>"]);
+                $parameterData['setter'] = fn() => $this->getFaker()->dateTimeBetween(null, $endDate)->format('Y-m-d');
+                break;
             case 'starts_with':
-                $parameterData['description'] .= 'The value must start with one of ' . w::getListOfValuesAsFriendlyHtmlString($arguments).' ';
+                $parameterData['description'] .= ' The value must start with one of ' . w::getListOfValuesAsFriendlyHtmlString($arguments);
                 $parameterData['setter'] = fn() => $this->getFaker()->lexify("{$arguments[0]}????");;
                 break;
             case 'ends_with':
-                $parameterData['description'] .= 'The value must end with one of ' . w::getListOfValuesAsFriendlyHtmlString($arguments).' ';
+                $parameterData['description'] .= ' The value must end with one of ' . w::getListOfValuesAsFriendlyHtmlString($arguments);
                 $parameterData['setter'] = fn() => $this->getFaker()->lexify("????{$arguments[0]}");;
                 break;
             case 'uuid':
-                $parameterData['description'] .= $this->getDescription($rule).' ';
+                $parameterData['description'] .= ' ' . $this->getDescription($rule) . ' ';
                 $parameterData['setter'] = fn() => $this->getFaker()->uuid();;
+                break;
+            case 'regex':
+                $parameterData['description'] .= ' ' . $this->getDescription($rule, [':regex' => $arguments[0]]);
+                $parameterData['setter'] = fn() => $this->getFaker()->regexify($arguments[0]);;
                 break;
 
             /**
              * Special number types. Some rules here may apply to other types, but we treat them as being numeric.
              */
             case 'digits':
-                $parameterData['description'] .= $this->getDescription($rule, [':digits' => $arguments[0]]) . ' ';
+                $parameterData['description'] .= ' ' . $this->getDescription($rule, [':digits' => $arguments[0]]);
                 $parameterData['setter'] = fn() => $this->getFaker()->randomNumber($arguments[0], true);
                 $parameterData['type'] = 'number';
                 break;
 
             case 'digits_between':
-                $parameterData['description'] .= $this->getDescription($rule, [':min' => $arguments[0], ':max' => $arguments[1]]) . ' ';
+                $parameterData['description'] .= ' ' . $this->getDescription($rule, [':min' => $arguments[0], ':max' => $arguments[1]]);
                 $parameterData['setter'] = fn() => $this->getFaker()->randomNumber($this->getFaker()->numberBetween(...$arguments), true);
                 $parameterData['type'] = 'number';
                 break;
-            /*
-             * min, max and between not supported until we can figure out a proper way
-             *  to make them compatible with multiple types (string, number, file)
+
+            /**
+             * These rules can apply to numbers, strings, arrays or files
+             */
+            case 'size':
+                $parameterData['description'] .= ' ' . $this->getDescription(
+                        $rule, [':size' => $arguments[0]], $this->getLaravelValidationBaseTypeMapping($parameterData['type'])
+                    );
+                $parameterData['setter'] = $this->getDummyValueGenerator($parameterData['type'], $arguments[0]);
+                break;
             case 'min':
-                $parameterData['type'] = $parameterData['type'] ?: 'number';
-                $parameterData['description'] .= Description::getDescription($rule, [':min' => $arguments[0]], 'numeric').' ';
-                $parameterData['setter'] = function () { return $this->getFaker()->numberBetween($arguments[0]); };
+                $parameterData['description'] .= ' ' . $this->getDescription(
+                        $rule, [':min' => $arguments[0]], $this->getLaravelValidationBaseTypeMapping($parameterData['type'])
+                    );
+                $parameterData['setter'] = $this->getDummyDataGeneratorBetween($parameterData['type'], $arguments[0]);
                 break;
             case 'max':
-                $parameterData['type'] = $parameterData['type'] ?: 'number';
-                $parameterData['description'] .= Description::getDescription($rule, [':max' => $arguments[0]], 'numeric').' ';
-                $parameterData['setter'] = function () { return $this->getFaker()->numberBetween(0, $arguments[0]); };
+                $parameterData['description'] .= ' ' . $this->getDescription(
+                        $rule, [':max' => $arguments[0]], $this->getLaravelValidationBaseTypeMapping($parameterData['type'])
+                    );
+                $parameterData['setter'] = $this->getDummyDataGeneratorBetween($parameterData['type'], 0, $arguments[0]);
                 break;
             case 'between':
-                $parameterData['type'] = $parameterData['type'] ?: 'number';
-                $parameterData['description'] .= Description::getDescription($rule, [':min' => $arguments[0], ':max' => $arguments[1]], 'numeric').' ';
-                $parameterData['setter'] = function () { return $this->getFaker()->numberBetween($arguments[0], $arguments[1]); };
-                break;*/
+                $parameterData['description'] .= ' ' . $this->getDescription(
+                        $rule, [':min' => $arguments[0], ':max' => $arguments[1]], $this->getLaravelValidationBaseTypeMapping($parameterData['type'])
+                    );
+                $parameterData['setter'] = $this->getDummyDataGeneratorBetween($parameterData['type'], $arguments[0], $arguments[1]);
+                break;
 
             /**
              * Special file types.
              */
             case 'image':
                 $parameterData['type'] = 'file';
-                $parameterData['description'] .= $this->getDescription($rule) . ' ';
+                $parameterData['description'] .= ' ' . $this->getDescription($rule) . ' ';
                 $parameterData['setter'] = function () {
                     // This is fine because the file example generator generates an image
                     return $this->generateDummyValue('file');
@@ -299,16 +369,58 @@ trait ParsesValidationRules
              */
             case 'in':
                 // Not using the rule description here because it only says "The attribute is invalid"
-                $parameterData['description'] .= 'The value must be one of ' . w::getListOfValuesAsFriendlyHtmlString($arguments) . ' ';
+                $parameterData['description'] .= ' The value must be one of ' . w::getListOfValuesAsFriendlyHtmlString($arguments) . ' ';
                 $parameterData['setter'] = function () use ($arguments) {
                     return Arr::random($arguments);
                 };
+                break;
+
+            /**
+             * These rules only add a description. Generating valid examples is too complex.
+             */
+            case 'not_in':
+                $parameterData['description'] .= ' The value must not be one of ' . w::getListOfValuesAsFriendlyHtmlString($arguments) . ' ';
+                break;
+            case 'required_if':
+                $parameterData['description'] .= ' ' . $this->getDescription(
+                        $rule, [':other' => "<code>{$arguments[0]}</code>", ':value' => "<code>{$arguments[1]}</code>"]
+                    ) . ' ';
+                break;
+            case 'required_unless':
+                $parameterData['description'] .= ' ' . $this->getDescription(
+                        $rule, [':other' => "<code>" . array_shift($arguments) . "</code>", ':values' => w::getListOfValuesAsFriendlyHtmlString($arguments)]
+                    ) . ' ';
+                break;
+            case 'required_with':
+                $parameterData['description'] .= ' ' . $this->getDescription(
+                        $rule, [':values' => w::getListOfValuesAsFriendlyHtmlString($arguments)]
+                    ) . ' ';
+                break;
+            case 'required_without':
+                $description = $this->getDescription(
+                        $rule, [':values' => w::getListOfValuesAsFriendlyHtmlString($arguments)]
+                    ) . ' ';
+                $parameterData['description'] .= str_replace('must be present', 'is not present', $description);
+                break;
+            case 'required_with_all':
+            case 'required_without_all':
+                $parameterData['description'] .= ' ' . $this->getDescription(
+                        $rule, [':values' => w::getListOfValuesAsFriendlyHtmlString($arguments, "and")]
+                    ) . ' ';
+                break;
+            case 'same':
+            case 'different':
+                $parameterData['description'] .= ' ' . $this->getDescription(
+                        $rule, [':other' => "<code>{$arguments[0]}</code>"]
+                    ) . ' ';
                 break;
 
             default:
                 // Other rules not supported
                 break;
         }
+
+        return true;
     }
 
     /**
@@ -334,7 +446,7 @@ trait ParsesValidationRules
         if (strpos($rule, ':') !== false) {
             [$rule, $argumentsString] = explode(':', $rule, 2);
 
-            // These rules can have ommas in their arguments, so we don't split on commas
+            // These rules can have commas in their arguments, so we don't split on commas
             if (in_array(strtolower($rule), ['regex', 'date', 'date_format'])) {
                 $ruleArguments = [$argumentsString];
             } else {
@@ -345,7 +457,7 @@ trait ParsesValidationRules
         return [strtolower(trim($rule)), $ruleArguments];
     }
 
-    protected function setParameterExampleAndDescription(array $parameterData, array $userSpecifiedParameterInfo): array
+    protected function updateParameterExample(array $parameterData): array
     {
         // If no example was given by the user, set an autogenerated example.
         // Each parsed rule returns a 'setter' function. We'll evaluate the last one.
@@ -353,19 +465,6 @@ trait ParsesValidationRules
             $parameterData['example'] = $parameterData['setter']();
         }
 
-        // Make sure the user-specified description comes first (and add full stops where needed).
-        $userSpecifiedDescription = $userSpecifiedParameterInfo['description'] ?? '';
-        if (!empty($userSpecifiedDescription) && !Str::endsWith($userSpecifiedDescription, '.')) {
-            $userSpecifiedDescription .= '.';
-        }
-        $validationDescription = trim($parameterData['description'] ?: '');
-        $fullDescription = trim($userSpecifiedDescription . ' ' . trim($validationDescription));
-        $parameterData['description'] = $fullDescription ? rtrim($fullDescription, '.') . '.' : $fullDescription;
-
-        // Set a default type
-        if (is_null($parameterData['type'])) {
-            $parameterData['type'] = 'string';
-        }
         // If the parameter is required and has no example, generate one.
         if ($parameterData['required'] === true && $parameterData['example'] === self::$MISSING_VALUE) {
             $parameterData['example'] = $this->generateDummyValue($parameterData['type']);
@@ -375,6 +474,7 @@ trait ParsesValidationRules
             // Casting again is important since values may have been cast to string in the validator
             $parameterData['example'] = $this->castToType($parameterData['example'], $parameterData['type']);
         }
+
         return $parameterData;
     }
 
@@ -504,7 +604,25 @@ trait ParsesValidationRules
             $description = str_replace($placeholder, $argument, $description);
         }
 
+        $description = str_replace("The :attribute field", "This field", $description);
         return str_replace(":attribute", "value", $description);
+    }
+
+    private function getLaravelValidationBaseTypeMapping(string $parameterType): string
+    {
+        $mapping = [
+            'number' => 'numeric',
+            'integer' => 'numeric',
+            'file' => 'file',
+            'string' => 'string',
+            'array' => 'array',
+        ];
+
+        if (Str::endsWith($parameterType, '[]')) {
+            return 'array';
+        }
+
+        return $mapping[$parameterType] ?? 'string';
     }
 
     protected function getMissingCustomDataMessage($parameterName)